refactor(Lotto): 검증 클래스를 내부로 이동
LottoValidator를 Lotto의 내부 클래스로 정의
